// js/src/common/types.ts
export interface Query {
  sql: string;
  preview: string;
  bindings: string[];
  duration: number;
  connection: string;
  origin: string | null;
  occurrences: number;
}

export interface QueriesData {
  count: number;
  dropped: number;
  duplicates: number;
  duration: number;
  queries: Query[];
}